requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiohttp @ file:///croot/aiohttp_1701112538292/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
appdirs==1.4.4
attrs @ file:///croot/attrs_1695717823297/work
auto_gptq==0.7.1
autoawq==0.2.8
autoawq_kernels==0.0.7
bcrypt==4.2.0
Bottleneck @ file:///work/ci_py311/bottleneck_1676827011612/work
Brotli @ file:///work/ci_py311/brotli-split_1676830125088/work
capstone==6.0.0a1
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chex==0.1.86
click==8.1.7
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1697464713350/work
colorama @ file:///work/ci_py311/colorama_1676827038675/work
colored-traceback==0.4.2
coloredlogs==15.0.1
commonmark @ file:///Users/ktietz/demo/mc3/conda-bld/commonmark_1630649545323/work
contourpy @ file:///croot/contourpy_1700583582875/work
cryptography @ file:///croot/cryptography_1702070282333/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython @ file:///home/conda/feedstock_root/build_artifacts/cython_1706274766471/work
datasets==3.2.0
dill @ file:///croot/dill_1692271232022/work
docker-pycreds==0.4.0
einops==0.8.0
etils==1.8.0
evaluate==0.4.3
filelock @ file:///croot/filelock_1700591183607/work
flash-attn==2.6.3
flax==0.8.3
fonttools==4.25.0
frozenlist @ file:///croot/frozenlist_1698702560391/work
fsspec @ file:///croot/fsspec_1701286474621/work
future @ file:///work/ci_py311_2/future_1679335928186/work
gekko==1.2.1
gitdb==4.0.11
GitPython==3.1.42
gmpy2 @ file:///work/ci_py311/gmpy2_1676839849213/work
grpcio==1.62.2
huggingface-hub==0.28.1
humanfriendly==10.0
idna @ file:///work/ci_py311/idna_1676822698822/work
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib_resources==6.4.0
intervaltree==3.1.0
jax==0.4.27
jaxlib @ https://storage.googleapis.com/jax-releases/cuda12/jaxlib-0.4.27+cuda12.cudnn89-cp311-cp311-manylinux2014_x86_64.whl#sha256=e59b367ca342edce5f478f2304d738284e11f5b98728f8fcbaf06164819b6456
Jinja2 @ file:///work/ci_py311/jinja2_1676823587943/work
joblib==1.3.2
kiwisolver @ file:///work/ci_py311/kiwisolver_1676827230232/work
loralib==0.1.2
Mako==1.3.6
Markdown==3.6
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1686175045316/work
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib @ file:///croot/matplotlib-suite_1698692105134/work
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1704317613764/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
ml-dtypes==0.3.2
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.0.8
multidict @ file:///croot/multidict_1701096859099/work
multiprocess @ file:///croot/multiprocess_1692294385131/work
munkres==1.1.4
Naked==0.1.32
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1690561992265/work
ninja==1.11.1.1
numexpr @ file:///croot/numexpr_1696515281613/work
numpy @ file:///croot/numpy_and_numpy_base_1704311704800/work/dist/numpy-1.26.3-cp311-cp311-linux_x86_64.whl#sha256=10a078151ecec16bafb535f7487635217625fa06536dec8509e514648c78d626
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver==11.1.1.58
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse==11.5.0.58
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvidia-pyindex==1.0.9
opt-einsum==3.3.0
optax==0.2.2
optimum==1.23.3
orbax-checkpoint==0.5.10
packaging @ file:///croot/packaging_1693575174725/work
pandas @ file:///croot/pandas_1702317985682/work/dist/pandas-2.1.4-cp311-cp311-linux_x86_64.whl#sha256=a15c481f210e34ff91dac4944a5f8f0fe38ff2c3afa7c45ea621ff460d6f9b57
paramiko==3.5.0
peft==0.14.0
Pillow @ file:///croot/pillow_1696580024257/work
platformdirs==4.3.6
plumbum==1.9.0
ply==3.11
protobuf==4.25.3
psutil==5.9.7
pwntools==4.13.1
pyarrow==19.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.21.0
pyelftools==0.31
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
PyNaCl==1.5.0
pynvml @ file:///home/conda/feedstock_root/build_artifacts/pynvml_1639061605391/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing @ file:///work/ci_py311/pyparsing_1677811559502/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
pyserial==3.5
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///croot/pytz_1695131579487/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
quanto==0.2.0
regex @ file:///croot/regex_1696515298636/work
requests==2.32.3
responses==0.18.0
rich @ file:///opt/conda/conda-bld/rich_1641600379360/work
ROPGadget==7.5
rouge==1.0.1
rpyc==6.0.1
safetensors==0.4.3
scalene @ file:///home/conda/feedstock_root/build_artifacts/scalene_1714185276361/work
scikit-learn==1.4.1.post1
scipy==1.12.0
sentencepiece==0.2.0
sentry-sdk==1.43.0
setproctitle==1.3.3
shellescape==3.8.1
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sortedcontainers==2.4.0
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorstore==0.1.58
threadpoolctl==3.4.0
tokenizers==0.21.0
toolz==0.12.1
torch==2.6.0
torch-tb-profiler==0.4.3
torchaudio==2.1.2
torchvision==0.16.2
tornado @ file:///croot/tornado_1696936946304/work
tqdm==4.66.4
transformers==4.47.1
triton==3.2.0
typing_extensions==4.12.2
tzdata @ file:///croot/python-tzdata_1690578112552/work
unicorn==2.1.1
unix-ar==0.2.1
urllib3 @ file:///croot/urllib3_1698257533958/work
wandb==0.18.3
Werkzeug==3.0.2
word2number==1.1
xxhash @ file:///work/ci_py311/python-xxhash_1676842384694/work
yarl @ file:///croot/yarl_1701105127787/work
zipp @ file:///croot/zipp_1704206909481/work
zstandard==0.23.0
